scriptorium/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
gunicorn==23.0.0
python-dotenv>=1.1.0
python-telegram-bot==21.7
pytesseract==0.3.13
Pillow>=11.2.1
Werkzeug==3.1.3
SQLAlchemy==2.0.36
openai==1.56.2
